+++ lib/src/cmaf/ham/types/mapper/dash/ContentComponent.ts
@@ -1,0 +1,6 @@
+export type ContentComponent = {
+	$: {
+		contentType: string;
+		id: string;
+	};
+};