+++ lib/src/cmaf/ham/types/mapper/hls/Playlist.ts
@@ -1,0 +1,18 @@
+/**
+ * HLS Playlist
+ *
+ * @group CMAF
+ * @alpha
+ */
+export type PlayList = {
+	uri: string;
+	attributes: {
+		FRAME_RATE: number;
+		CODECS: string;
+		BANDWIDTH: number;
+		RESOLUTION: {
+			width: number;
+			height: number;
+		};
+	};
+};