+++ lib/src/cmaf/ham/types/model/Presentation.ts
@@ -1,0 +1,14 @@
+import { Ham } from './Ham.js';
+import { SelectionSet } from './SelectionSet.js';
+
+/**
+ * CMAF-HAM Presentation type
+ * 
+ * selectionSets - List of SelectionSets
+ * 
+ * @group CMAF
+ * @alpha
+ */
+export type Presentation = Ham & {
+	selectionSets: SelectionSet[];
+};